+++ NEWS.md
@@ -1,3 +1,10 @@
+
+# hyperSpec.utils 0.0.0.9001
+
+* homogenized arguments and output of unmixing and mixture analysis functions (backwards incompatible changes!)
+* fixed error in `nmf()`
+* added `omp()` and `ica()`
+
 # hyperSpec.utils 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.